pyproject.toml
[tool.poetry]
name = "tz-mbdigital"
version = "0.1.0"
description = ""
authors = ["Nazar <filoniuk.nazar.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
django-cors-headers = "^4.4.0"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
djoser = "^2.2.3"
django-crum = "^0.7.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
